Revert "Merge 7bb1151ba21e26d91ddaa83177bb58b4d1c36710 for LLVM update to 468909"
This reverts commit 8f38a8208f46ffc1bb0f83ebc754aae0a06589da.

Revert to make merge_from_upstream.py work.

Bug: 253033919
Test: presubmit

+++ compiler-rt/lib/sanitizer_common/symbolizer/sanitizer_symbolize.cpp
@@ -10,7 +10,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include <inttypes.h>
 #include <stdio.h>
 
 #include <string>
@@ -42,7 +41,7 @@ static llvm::symbolize::PrinterConfig getDefaultPrinterConfig() {
 }
 
 namespace __sanitizer {
-int internal_snprintf(char *buffer, uintptr_t length, const char *format,
+int internal_snprintf(char *buffer, unsigned long length, const char *format,
                       ...);
 }  // namespace __sanitizer
